add brand module

+++ src/brands/brands.service.ts
@@ -1,17 +1,27 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import {
+  Model,
+  PaginateModel,
+  PaginateOptions,
+  isValidObjectId,
+} from 'mongoose';
 import { Brand } from './brands.schema';
 import { CreateBrandDto } from './dto/create-brand.dto';
 import { GetBrandsQueryDto } from './dto/get-brand-query.dto';
 import { UpdateBrandDto } from './dto/update-brand.dto';
-
+import { ImagesService } from 'src/images/images.service';
 
 @Injectable()
 export class BrandsService {
   //
   constructor(
-    @InjectModel(Brand.name) private brandModel: Model<Brand>,
+    @InjectModel(Brand.name) private brandModel: PaginateModel<Brand>,
+    private imageService: ImagesService,
   ) {}
   //
   createBrand(createBrandData: CreateBrandDto) {
@@ -19,19 +29,51 @@ export class BrandsService {
     return newBrand.save();
   }
 
-  deleteBrand(brandId: string) {
-    return this.brandModel.findByIdAndDelete(brandId);
+  async deleteBrand(brandId: string) {
+    const brand = await this.brandModel.findById(brandId);
+    debugger;
+    if (!brand)
+      throw new NotFoundException(`no brand found with id ${brandId}`);
+    await this.imageService.deleteImage(brand.image as unknown as string);
+    return await this.brandModel.findByIdAndDelete(brandId);
   }
 
-  updateBrand(id: string, updateBrandData: UpdateBrandDto) {
-    return this.brandModel.findByIdAndUpdate(id, updateBrandData, {
-      runValidators: true,
-      new: true,
-    });
+  async updateBrand(id: string, updateBrandData: UpdateBrandDto) {
+    debugger;
+    if (!updateBrandData.image) {
+      const newBrand = await this.brandModel.findByIdAndUpdate(
+        id,
+        updateBrandData,
+        {
+          runValidators: true,
+          new: true,
+        },
+      );
+      if (!newBrand)
+        throw new NotFoundException(`no brand found with id ${id}`);
+      debugger;
+      return newBrand;
+    }
+    // 90881
+
+    const oldBrand = await this.brandModel.findById(id);
+    if (!oldBrand) throw new NotFoundException(`no brand found with id ${id}`);
+
+    const newBrand = await this.brandModel.findByIdAndUpdate(
+      id,
+      updateBrandData,
+      { runValidators: true, new: true },
+    );
+    debugger;
+    await this.imageService.deleteImage(oldBrand.image as unknown as string);
+    return newBrand;
   }
 
-  getBrandList(paginateOptions?: GetBrandsQueryDto) {
-    return (this.brandModel as any).paginate({}, paginateOptions);
+  getBrandList(paginateOptions?: PaginateOptions) {
+    return this.brandModel.paginate(
+      {},
+      { ...paginateOptions, populate: 'image' },
+    );
   }
   //
 }
